+++ Changelog.md
@@ -2,7 +2,8 @@
 
 ### 1.3 (2022-10-13)
 
-- `create_file()` and `create_dir()` now also work with Path instances
+- `create_file()`, `create_dir()`, `add_real_file()`, and `add_real_directory()`
+  now also work with Path instances
 - fixed `FakeFS.add_real_path()`
 - fix requirement of pyfakefs when using Python 2